Ifs agora funcionam
E o ParseBlock funciona agora como deveria desde o começo
