+++ test.lua
@@ -1,3 +1,9 @@
 asdf = "foda-se"
 
-print( (2 * 5) + 32 - 22 )
+if (2 = 5) then
+	print("Foi de prima")
+elseif (2 = 2) then
+	print("ta no elseif, krl")
+else
+	print("Saporra entrou no else")
+end
